implement budgets

+++ internal/models/budget.go
@@ -7,7 +7,7 @@ type Budget struct {
 	CategoryID int       `json:"category_id"`
 	TimePeriod int       `json:"time_period"`
 	UserID     int       `json:"user_id"`
-	Amount     float64   `json:"amount"`
+	Amount     int       `json:"amount"`
 	CreatedAt  time.Time `json:"created_at"`
 	UpdatedAt  time.Time `json:"updated_at"`
 }
